+++ tree/__init__.py
@@ -1,0 +1,10 @@
+from .tree import Tree
+from .tree_deck import TreeDeck
+
+tree_deck = TreeDeck()
+
+__all__ = [
+    'Tree',
+    'TreeDeck',
+    'tree_deck',
+]
