+++ tree/tree_deck.py
@@ -7,7 +7,7 @@ from typing import List, Dict
 
 from pandas import read_csv
 
-from _tree import Tree
+from tree import Tree
 
 
 class TreeDeck:
@@ -17,7 +17,7 @@ class TreeDeck:
 
     @staticmethod
     def _import_from_csv() -> List['Tree']:
-        df = read_csv("data/trees.csv")
+        df = read_csv('data/trees.csv')
         trees = df.values.tolist()
         return list(map(lambda tree: Tree(*tree), trees))
 
@@ -25,9 +25,8 @@ class TreeDeck:
         return self.cards.pop()
 
 
-tree_deck = TreeDeck()
-
 if __name__ == '__main__':
+    tree_deck = TreeDeck()
     tree_name_lengths: Dict[int, Tree] = {len(tree.common_name): tree for tree in tree_deck.cards}
     longest_name_length: int = max(tree_name_lengths.keys())
